+++ src/Symfony/Bundle/WebProfilerBundle/CHANGELOG.md
@@ -4,7 +4,9 @@ CHANGELOG
 4.4.0
 -----
 
-* Added button to clear the ajax request tab
+ * Added button to clear the ajax request tab
+ * Deprecated the `ExceptionController::templateExists()` method
+ * Deprecated the `TemplateManager::templateExists()` method
 
 4.3.0
 -----